pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Denis Kazakov <dykazakov@avito.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
mypy = "^1.10.1"
pytest = "^8.2.2"
types-python-dateutil = "^2.9.0.20240316"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
